ansi_str

Trait AnsiStr

pub trait AnsiStr {

    // Required methods
    fn ansi_get<I>(&self, i: I) -> Option<Cow<'_, str>>
       where I: RangeBounds<usize>;
    fn ansi_cut<I>(&self, i: I) -> Cow<'_, str>
       where I: RangeBounds<usize>;
    fn ansi_is_char_boundary(&self, index: usize) -> bool;
    fn ansi_find(&self, pat: &str) -> Option<usize>;
    fn ansi_strip_prefix(&self, prefix: &str) -> Option<Cow<'_, str>>;
    fn ansi_strip_suffix(&self, pat: &str) -> Option<Cow<'_, str>>;
    fn ansi_split<'a>(&'a self, pat: &'a str) -> AnsiSplit<'a> ;
    fn ansi_split_at(&self, mid: usize) -> (Cow<'_, str>, Cow<'_, str>);
    fn ansi_starts_with(&self, pat: &str) -> bool;
    fn ansi_ends_with(&self, pat: &str) -> bool;
    fn ansi_trim(&self) -> Cow<'_, str>;
    fn ansi_strip(&self) -> Cow<'_, str>;
    fn ansi_has_any(&self) -> bool;

}

AnsiStr represents a list of functions to work with colored strings defined as ANSI control sequences.

Required Methods

fn ansi_get<I>(&self, i: I) -> Option<Cow<'_, str>>
where I: RangeBounds<usize>,

Returns a substring of a string.

It preserves accurate style of a substring.

Range is defined in terms of bytes of the string not containing ANSI control sequences (If the string is stripped).

This is the non-panicking alternative to [Self::ansi_cut]. Returns None whenever equivalent indexing operation would panic.

Exceeding the boundaries of the string results in the same result if the upper boundary to be equal to the string length.

If the text doesn’t contains any ansi sequences the function must return result if [str::get] was called.

Examples

Basic usage:

use ansi_str::AnsiStr;

let text = "\u{1b}[31m🗻 on the 🌏\u{1b}[39m";

assert_eq!(text.ansi_get(0..7), Some("\u{1b}[31m🗻 on\u{1b}[39m".into()));

// indices not on UTF-8 sequence boundaries
assert!(text.ansi_get(1..).is_none());
assert!(text.ansi_get(..13).is_none());

// going over boundries doesn't panic
assert!(text.ansi_get(..std::usize::MAX).is_some());
assert!(text.ansi_get(std::usize::MAX..).is_some());

Text doesn’t contain ansi sequences

use ansi_str::AnsiStr;

let text = "🗻 on the 🌏";

assert_eq!(text.ansi_get(5..), Some("on the 🌏".into()));

fn ansi_cut<I>(&self, i: I) -> Cow<'_, str>
where I: RangeBounds<usize>,

Cut makes a sub string, keeping the colors in the substring.

The ANSI escape sequences are ignored when calculating the positions within the string.

Range is defined in terms of bytes of the string not containing ANSI control sequences (If the string is stripped).

Exceeding an upper bound does not panic.

Panics

Panics if a start or end indexes are not on a UTF-8 code point boundary.

Examples

Basic usage:

use ansi_str::AnsiStr;

let text = "\u{1b}[31;40m🗻 on the 🌏\u{1b}[0m";
assert_eq!(text.ansi_cut(0..4).ansi_strip(), "🗻");
assert_eq!(text.ansi_cut(..7).ansi_strip(), "🗻 on");
assert_eq!(text.ansi_cut(8..).ansi_strip(), "the 🌏");

Panics when index is not a valud UTF-8 char

use ansi_str::AnsiStr;

let text = "\u{1b}[31;40m🗻 on the 🌏\u{1b}[0m";
text.ansi_cut(1..);

fn ansi_is_char_boundary(&self, index: usize) -> bool

Checks that index-th byte is the first byte in a UTF-8 code point sequence or the end of the string.

The index is determined in a string if it would be stripped.

Examples

Basic usage:

use ansi_str::AnsiStr;

let text = "\u{1b}[34mLöwe 老虎 Léopard\u{1b}[39m";

assert!(text.ansi_is_char_boundary(0));
// start of `老`
assert!(text.ansi_is_char_boundary(6));
assert!(text.ansi_is_char_boundary(text.ansi_strip().len()));

// second byte of `ö`
assert!(!text.ansi_is_char_boundary(2));

// third byte of `老`
assert!(!text.ansi_is_char_boundary(8));

fn ansi_find(&self, pat: &str) -> Option<usize>

Returns the byte index of the first character of this string slice that matches the pattern, considering the ansi sequences.

Returns None if the pattern doesn’t match.

Examples

Basic usage:

use ansi_str::AnsiStr;

let text = "\u{1b}[31;40mLöwe 老虎 Léopard Gepardi\u{1b}[0m";
assert_eq!(text.ansi_find("L"), Some(0));
assert_eq!(text.ansi_find("é"), Some(14));
assert_eq!(text.ansi_find("pard"), Some(17));

fn ansi_strip_prefix(&self, prefix: &str) -> Option<Cow<'_, str>>

Returns a string with the prefix removed, considering the ansi sequences.

If the string starts with the pattern prefix, returns substring after the prefix, wrapped in Some.

If the string does not start with prefix, returns None.

Examples

Basic usage:

use ansi_str::AnsiStr;

let text = "\u{1b}[31mfoo:bar\u{1b}[0m";
assert_eq!(
    text.ansi_strip_prefix("foo"),
    Some("\u{1b}[31m:bar\u{1b}[0m".into()),
);
assert_eq!(
    text.ansi_strip_prefix("bar"),
    None,
);

fn ansi_strip_suffix(&self, pat: &str) -> Option<Cow<'_, str>>

Returns a string slice with the suffix removed, considering the ansi sequences.

If the string ends with the pattern suffix, returns the substring before the suffix, wrapped in Some.

If the string does not end with suffix, returns None.

Examples

Basic usage:

use ansi_str::AnsiStr;

let text = "\u{1b}[31mfoo:bar\u{1b}[0m";
assert_eq!(text.ansi_strip_suffix("bar"), Some("\u{1b}[31mfoo:\u{1b}[0m".into()));
assert_eq!(text.ansi_strip_suffix("foo"), None);

fn ansi_split<'a>(&'a self, pat: &'a str) -> AnsiSplit<'a>

An iterator over substrings of the string, separated by characters matched by a pattern. While keeping colors in substrings.

Examples

Basic usage:

use ansi_str::AnsiStr;

let text = "\u{1b}[31mMary had a little lamb\u{1b}[0m";

let words: Vec<_> = text.ansi_split(" ").collect();

assert_eq!(
    words,
    [
        "\u{1b}[31mMary\u{1b}[39m",
        "\u{1b}[31mhad\u{1b}[39m",
        "\u{1b}[31ma\u{1b}[39m",
        "\u{1b}[31mlittle\u{1b}[39m",
        "\u{1b}[31mlamb\u{1b}[0m",
    ]
);

let words: Vec<_> = "".ansi_split("X").collect();
assert_eq!(words, [""]);

let text = "\u{1b}[31mlionXXtigerXleopard\u{1b}[0m";
let words: Vec<_> = text.ansi_split("X").collect();
assert_eq!(words, ["\u{1b}[31mlion\u{1b}[39m", "", "\u{1b}[31mtiger\u{1b}[39m", "\u{1b}[31mleopard\u{1b}[0m"]);

let text = "\u{1b}[31mlion::tiger::leopard\u{1b}[0m";
let words: Vec<_> = text.ansi_split("::").collect();
assert_eq!(words, ["\u{1b}[31mlion\u{1b}[39m", "\u{1b}[31mtiger\u{1b}[39m", "\u{1b}[31mleopard\u{1b}[0m"]);

fn ansi_split_at(&self, mid: usize) -> (Cow<'_, str>, Cow<'_, str>)

Divide one string into two at an index. While considering colors.

The argument, mid, should be a byte offset from the start of the string. It must also be on the boundary of a UTF-8 code point.

The two strings returned go from the start of the string to mid, and from mid to the end of the string.

Panics

It might panic in case mid is not on the boundry of a UTF-8 code point.

Examples

Basic usage:

use ansi_str::AnsiStr;

let text = "\u{1b}[31;40mPer Martin-Löf\u{1b}[0m";

let (first, last) = text.ansi_split_at(3);

assert_eq!(first.ansi_strip(), "Per");
assert_eq!(last.ansi_strip(), " Martin-Löf");

Panic

use ansi_str::AnsiStr;

let text = "\u{1b}[31;40mPer Martin-Löf\u{1b}[0m";

text.ansi_split_at(13);

fn ansi_starts_with(&self, pat: &str) -> bool

Returns true if the given pattern matches a prefix of this string slice. Ignoring the ansi sequences.

Returns false if it does not.

Examples

Basic usage:

use ansi_str::AnsiStr;

let text = "\u{1b}[31;40mbananas\u{1b}[0m";

assert!(text.ansi_starts_with("bana"));
assert!(!text.ansi_starts_with("nana"));

fn ansi_ends_with(&self, pat: &str) -> bool

Returns true if the given pattern matches a suffix of this string slice. Ignoring the ansi sequences.

Returns false if it does not.

Examples

Basic usage:

use ansi_str::AnsiStr;

let text = "\u{1b}[31;40mbananas\u{1b}[0m";

assert!(text.ansi_ends_with("anas"));
assert!(!text.ansi_ends_with("nana"));

fn ansi_trim(&self) -> Cow<'_, str>

Returns a string slice with leading and trailing whitespace removed. Ignoring the ansi sequences.

Examples

Basic usage:

use ansi_str::AnsiStr;

let text = String::from("\u{1b}[31m Hello\tworld\t\u{1b}[39m");

assert_eq!(text.ansi_trim(), "\u{1b}[31mHello\tworld\u{1b}[39m");

fn ansi_strip(&self) -> Cow<'_, str>

Returns a string with all ANSI sequences removed.

Examples

Basic usage:

use ansi_str::AnsiStr;

let text = "\u{1b}[31;40mHello World!\u{1b}[0m";

assert_eq!(text.ansi_strip(), "Hello World!");

fn ansi_has_any(&self) -> bool

Returns true if a string contains any ansi sequences.

Returns false if it does not.

Examples

Basic usage:

use ansi_str::AnsiStr;

assert!(!"Hi".ansi_has_any());
assert!("\u{1b}[31;40mHi\u{1b}[0m".ansi_has_any());

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.